Image-Classification-yolo/requirements.txt
ultralytics>=8.0.0
fastapi
uvicorn
python-multipart
Pillow
